refactor(calendar): extract GoogleCalendarIntegration from settings menu
